Fixing the query parameter

+++ BackEndFlask/controller/Routes/Student_bulk_upload_routes.py
@@ -24,36 +24,41 @@ def upload_CSV():
         print("[UploadCsv_routes /upload POST] Unsuccessfully uploaded a .csv file! Wrong Format")
         createBadResponse("Unsuccessfully uploaded a .csv or .xlsx file!", "Wrong Format","students")
         return response
-    try:
-        directory = os.path.join(os.getcwd(), "Test")
-        os.makedirs(directory, exist_ok=True)
-        file_path = os.path.join(directory, file.filename)
-        file.save(file_path)
-        result = studentImport.studentcsvToDB(file_path,1,1)
-
-        if isinstance(result, str):
-            shutil.rmtree(directory)
-            print(f"[UploadCsv_routes /upload POST] Unsuccessfully uploaded a {extension[1]} file! Error Raised!")
-            createBadResponse(f"Unsuccessfully uploaded a {extension[1]} file!", str(result),"students")
-            return response
-        shutil.rmtree(directory)
-        
-        file.seek(0,0)
-        file_data = file.read()
-        if extension[1] == ".csv":
-            df = pd.read_csv(BytesIO(file_data))
-        else:
-            df = pd.read_excel(BytesIO(file_data))
-        
-        headers = df.columns
-        df.columns = ["Student","lms_id", "email"]
-        df.loc[len(df.index)] = headers
-        results = json.loads(df.to_json(orient="records"))
-        file.seek(0,0)
-            
-        print(f"[UploadCsv_routes /upload POST] Successfully uploaded a {extension[1]} file!")
-        createGoodResponse(f"Successfully uploaded a {extension[1]} file!",results,200,"students")
-        return response
     
-    except Exception:
-        pass
+    if request.args.get("course_id"):
+        course_id = int(request.args.get("course_id"))
+        try:
+            directory = os.path.join(os.getcwd(), "Test")
+            os.makedirs(directory, exist_ok=True)
+            file_path = os.path.join(directory, file.filename)
+            file.save(file_path)
+            result = studentImport.studentcsvToDB(file_path,1,course_id)
+
+            if isinstance(result, str):
+                shutil.rmtree(directory)
+                print(f"[UploadCsv_routes /upload POST] Unsuccessfully uploaded a {extension[1]} file! Error Raised!")
+                createBadResponse(f"Unsuccessfully uploaded a {extension[1]} file!", str(result),"students")
+                return response
+            shutil.rmtree(directory)
+            
+            file.seek(0,0)
+            file_data = file.read()
+            if extension[1] == ".csv":
+                df = pd.read_csv(BytesIO(file_data))
+            else:
+                df = pd.read_excel(BytesIO(file_data))
+            
+            headers = df.columns
+            df.columns = ["Student","lms_id", "email"]
+            df.loc[len(df.index)] = headers
+            results = json.loads(df.to_json(orient="records"))
+            file.seek(0,0)
+                
+            print(f"[UploadCsv_routes /upload POST] Successfully uploaded a {extension[1]} file!")
+            createGoodResponse(f"Successfully uploaded a {extension[1]} file!",results,200,"students")
+            return response
+    
+        except Exception:
+            pass
+    else:
+        createBadResponse(f"Unsuccessfully uploaded a file!", "Course_id was not passed.","students")
